# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Sample API",
        "contact": {
            "name": "Erik Alexander, Friscilia Sutanto, Sri Wardani",
            "url": "https://uph.edu"
        },
        "version": "1.0"
    },
    "host": "webservice.app",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/v1/jurusan": {
            "get": {
                "tags": [
                    "Jurusan"
                ],
                "summary": "Retrieves the collection of Jurusan resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jurusan collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/jurusan"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "post": {
                "tags": [
                    "Jurusan"
                ],
                "summary": "Data Jurusan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Data Jurusan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users Token.",
                        "schema": {
                            "property": "token",
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/v1/jurusan/{id}": {
            "get": {
                "tags": [
                    "Jurusan"
                ],
                "summary": "Retrieves the collection of Jurusan resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jurusan collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/jurusan"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "put": {
                "tags": [
                    "Jurusan"
                ],
                "summary": "Data Jurusan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Data Jurusan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Update.",
                        "schema": {
                            "property": "token",
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Jurusan"
                ],
                "summary": "Removes the Jurusan resource.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Jurusan resource delete."
                    },
                    "401": {
                        "description": "Unauthorized Action."
                    },
                    "404": {
                        "description": "Resource not found."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "jurusan": {
            "properties": {
                "nama": {
                    "type": "string"
                },
                "id_fakultas": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        },
...
    }
}